Fixed issue with listening for HEX ints vs DECIMAL ints and LONG ints

+++ src/edu/mit/compilers/ir/IrLiteralInt.java
@@ -1,9 +1,9 @@
 package edu.mit.compilers.ir;
 
 public class IrLiteralInt extends IrLiteral {
-    private final int value;
+    private final long value;
 
-    public IrLiteralInt(int value, int lineNumber, int colNumber) {
+    public IrLiteralInt(long value, int lineNumber, int colNumber) {
         super(lineNumber, colNumber);
         this.value = value;
     }
